refactor: implement PublicWorkspaceRoute for unified public bootstrapping
